Add zip::ZipFiles() function which creates a zip file for a specified list of input files.
Also added unit test for ZipFiles().

This is to permit caller to specify an arbitrary subset of file contents under a directory to
be included in the created zip file. Although it is possible to achieve this with the existing
function ZipWithFilterCallback(), the required filter callback function would be quite messy
and inefficient.

BUG=138359
TEST=unit_tests, manual tests

Review URL: https://chromiumcodereview.appspot.com/11392005