Merge 166509 - touch: Do not crash if touch-handlers are uninstalled in response to a touch-event.
If a touch-event handler removes all the touch event handlers on the page
in response to a touch-event, then the touch-queue for that page will have
been flushed by the time the ACK for the event comes in. Handle such cases
gracefully to avoid a crash.

BUG=157783

Review URL: https://codereview.chromium.org/11378002

TBR=sadrul@chromium.org
Review URL: https://codereview.chromium.org/11359145