Have ui_unittest_strings target depend on ui_strings target.
The ui_unittest_strings target assumes the presence of files that are built by
the ui_strings target, but the former wasn't depending on the latter. The fact
that this dependence didn't exist could cause builds of ui_unittests to fail
flakily on platforms for which ui_unittests depends on ui_unittest_strings.


Review URL: https://chromiumcodereview.appspot.com/11360167