Do not build 'all' on all the try builders by default.
Enforce the parameters options and tests to make sure it's not forgotten.
Significantly reduce the number of unneeded targets that are built.
This will greatly improve the TS performance.

The main reason for this CL is that foo_tests_run are not built by default.
These target archive the build output, so they must not be built unless
necessary. There is already logic to only build the foo_test_run for each
foo_test_swarm testfilter, but that is not useful as long as 'all' is built.

TBR=cmp@chromium.org
R=petermayo@chromium.org,glider@chromium.org
BUG=


Review URL: https://chromiumcodereview.appspot.com/11365154