Set root resource ID upon full feed update.
We used to set the resource ID at DriveResourceMetadata's 
constructor as for WAPI the resource ID for root is 
a fixed string, but for Drive API it is not fixed and 
was set when About resource is obtained. 

This caused a trouble for maintaining the code and 
this CL is trying to consolidate setting root resource 
ID at one place. 

Although for Drive API 'root' can be used for a convenient 
alias for queries, Drive API usually returns a reference 
to root directory (e.g. reference to parent folder) 
as its real root directory resource ID rather than 'root' 
alias, therefore we don't use it here. 

This was originally reviewed in https://chromiumcodereview.appspot.com/11227020/
and submitted three times (r164876, r165219, r165303) but reverted due to
unit_test flakiness (DriveFileSystemTest.ContentSearchWithNewEntry failed
once in a while indeterministically).

Removed parts that uses a different root resource ID for unit tests.

BUG=152230, 152288
TEST=compiles, pass all existing tests.


Review URL: https://chromiumcodereview.appspot.com/11361115