ui: Use Vector2d in the View::CalculateOffsetToAncestorWithLayer API
This function computes an offset, so it should be a vector. Since we have
math operations now, instead of passing in a Point*, we return a new
Vector2d, which we can add to other Vector2ds, or use as is.

Depends on https://codereview.chromium.org/11293194/

R=pkasting
BUG=158615

Review URL: https://codereview.chromium.org/11362179