Remove a DCHECK() in FieldTrial::SetForced().
The DCHECK() was added under the assumption that it doesn't
make sense to force a trial that only has the default group.

However, with VariationsService, which is general-purpose,
one could create a server-side trial of this sort, and this
works correctly other than the DCHECK().

BUG=160766
TEST=New unit tests in field_trial_unittest.cc.
Review URL: https://codereview.chromium.org/11366205