cc: Add command line flags for the various show rect debugging features.
Add command line flags for everything except showPaintRects, as this is
already a part of WebSettings and passed in via WebKit. This needs to be
resolved with the others once we sort out the future method of passing in
settings to the compositor.

The new flags are:
--show-property-changed-rects
--show-surface-damage-rects
--show-screenspace-rects
--show-replica-screenspace-rects
--show-occluding-rects

BUG=160643
R=enne,jamesr

Review URL: https://codereview.chromium.org/11366212