Fix the crash in DesktopBackgroundView::OnPaint
ImageSkia is not RefCountedThreadSafe, just RefCounted. We are accessing wallpaper(ImageSkia) on different threads(FILE and UI). When two threads try to AddRef at the same time, it cause problem. For m23, the current solution is to make a deep copy of imageskia and pass a scoped pointer to other thread. See (http://codereview.chromium.org/11361131/)

BUG=159431

Review URL: https://chromiumcodereview.appspot.com/11366143