cc: Layer that have empty bounds should still have their contentsScale applied in their transforms.
The change to add contentsScaleX() and contentsScaleY() allows layers to move
between their layer and content space even when they have empty bounds. Now new
code is moving between these spaces by multiplying these scale factors,
regardless of the layer's bounds, and this seems like the correct thing to do.

However, LayerTreeHostCommon is still only applying contentsScaleX|Y() only when
the bounds are not empty. This is inconsistent with other code, and ends up
causing confusion elsewhere, such as in the scrolling and hit testing code.

We remove the if() conditions on the bounds in LTHCommon, so that the transforms
on any layer match the contentsScales on that layer.

Tests:
cc_unittests:LayerTreeHostCommonTest.verifyLayerTransformsInHighDPI
cc_unittests:LayerTreeHostCommonTest.verifyContentsScale

BUG=160834
R=enne


Review URL: https://chromiumcodereview.appspot.com/11312226