Revert 167794 - Native remoting host configurer for linux.
This is initial version of the remoting host configurer for linux. It doesn't look pretty yet,
but it works.

This requires a new dependency and a PSA was not sent in advance to chromium-dev.  It breaks
out bots because someone needs to manually run install-build-deps on each of them.

Please wait a couple days before landing.  yfriedman has sent the required PSA.


Review URL: https://chromiumcodereview.appspot.com/11260004

TBR=sergeyu@chromium.org
Review URL: https://codereview.chromium.org/11362266