cc: Visualize non-occluding rects in composited layers.
Adds a new flag to show non-occluding rects. When this flag is on, then for any
layer which could have been occluding, but isn't we show a rectangle on the
screen.

This means for layers that have 3d transforms and would never be occluding, we
don't draw rectangles. For axis aligned rectangles with opacity < 1, we don't
draw rectangles.

But when a layer has opacity=1, is axis aligned, but its contents are not
considered opaque, then we draw a rect to warn the user that this area could
have been occluding if only its contents had been considered opaque.

New command-line flag:
--show-nonoccluding-rects

R=jamesr,enne
BUG=160643

Review URL: https://codereview.chromium.org/11377122