Fix crash when programatically closing media gallery dialog
If the media gallery dialog was closed by something other than clicking "Cancel" or "Ok" then Chrome would crash.

The problem was that a bug in the way that I was closing the dialog. I was doing endSheet:alert when I should be doing endSheet:[alert window].

BUG=156537
TBR=ben@chromium.org

Review URL: https://chromiumcodereview.appspot.com/11369253