Add DecryptingDemuxerStream.
DecryptingDemuxerStream transforms an encrypted demuxer stream into a clear demuxer stream with the same stream type (audio/video) and codec config. It may be blocking because it relies on a Decryptor for decryption. Thus other than implementing the DemuxerStream interface, it also exposes a Reset() methods so that caller can cancel the decryption operation.

BUG=123421,141786
TEST=media_unittests pass

Review URL: https://chromiumcodereview.appspot.com/11342031