Keep playing audio even when opening another tab, going to another app etc.
Test: run both content shell and chrome for android, gone to a few html5 serving pages.

http://romaindillet.github.com/That-Radio-Player/#
and
http://grooveshark.com/#!/playlist/ac-dc/28046570

Then verified that audio keeps playing if I switch tabs or leave the app.

Also added some logging to verify that no matter what the internal player is released when you leave either when you leave the page if nothing is playing or upon closing the tab or navigating away if it is.




BUG=121898

Review URL: https://chromiumcodereview.appspot.com/11361239