Make runtest.py's --run-python-script argument actually work.
The runtest script checks that the given test_exe lives in the
build directory. This broke --run-python-script because python scripts
do not live in the build dir.

There was only one user of --run-python-script, which only used it on
Windows, which worked because the build-dir and target was given as ''.
For the other platforms, the _GetPythonTestCommand function worked around
the problem. This function has been removed.
Review URL: https://codereview.chromium.org/11293200