Remove top and bottom margins from TrayBubbleView
This CL modifies TrayBubbleView so that instead of ~3px margins filled
with a solid color at its top and bottom, the contents extends all the
way to the edges. To prevent the content from drawing over the bubble's
rounded corners, a mask is set on the hosting widget.

Since the masking would not work on Windows non-aura, the previous
behavior is preserved there.

BUG=159682


Review URL: https://chromiumcodereview.appspot.com/11293124