Ensure that the foreground tab's process has high priority
To reclaim memory, Android will kill renderer processes before killing
the browser process (See
http://developer.android.com/guide/components/processes-and-threads.html).
That is, the browser process is a foreground process while the renderer
processes are service processes. A foreground process can bind to a
service process so that that process has the same priority as the
foreground process (in terms of what should be killed).

BUG=159980


Review URL: https://chromiumcodereview.appspot.com/11361252