Implement the Flash menu as a new-style Pepper resource.
The code is now moved out of content and into Chrome. The functionality should be the same. Some of the ways that the host interacted with the PluginInstance (i.e. getting the fullscreen container) was tricky so I unfortunately had to add some functions to the RendererPpapiHost to expose this.

Review URL: https://codereview.chromium.org/11369176