Specify jar name for apks in java_apk.gypi
Previously, we had to keep this name in sync in multiple places, the
"fake" xxx_java target and in sdk-targets.xml. Instead, it should be set
only in java_apk.gypi and then passed along to each of the places that
we were specifying it before.

Also, delete unused 'output' block and unneeded xxx_java dependencies.

BUG=


Review URL: https://chromiumcodereview.appspot.com/11363131