Plug IOSurface leak at context tear-down.
Change the ordering of the destruction of a command buffer's
surface and context, so that the context is current while the
surface is destroyed.

Add lots of asserts to the surface tear-down path to verify
that we have the context that we need in order to delete the
resources.

If we don't delete the FBO and texture, then they will hold a
reference to the IOSurface until the share group is destroyed,
which, in the case of tabs opened from Gmail, may be a very
long time in the future.

BUG=158469


Review URL: https://chromiumcodereview.appspot.com/11418015