Add a class to replace ImageLoadingTracker with a nicer API.
This adds a new extensions::ImageLoader class, which mostly serves the same purpose as ImageLoadingTracker, but with some differences. It currently doesn't do any caching, and it uses callbacks instead of a delegate to pass the loaded image back to the caller. This should make this new class much nicer to use.
Also ported some ILT usage to the new ImageLoader class.

BUG=141673


Review URL: https://chromiumcodereview.appspot.com/11027044