Allow the --install-from-webstore flag to work when chrome is already running.
This requires the flag to be handled later in the startup sequence, at a point that is executed for cold startups, and also for chrome handing control over to an existing, running process.

This change also means chrome does not return a status code signifying the success or failure of the operation.

BUG=160607


Review URL: https://chromiumcodereview.appspot.com/11413004