browser_tests.isolate on Windows: Move all touched files from linux to !mac.
This is because the Windows tracer doesn't trace "touched files";
e.g. a file that is touched but not opened or a file that was empty.
There is not much cost to depend on then even if not necessary, since
their content is not needed.

TBR=csharp@chromium.org
NOTRY=true
BUG=


Review URL: https://chromiumcodereview.appspot.com/11299050