Android: SurfaceTexture-based image transport
This is an interim solution for where the embedder cannot use a command buffer context running on the gpu thread for frontend rendering (i.e. for the browser compositor), for example draw functor/callback based UI drawing.
On Android, sharing GL resources across threads is not a given and SurfaceTexture is therefore the only compatible option.

TBR=jam@chromium.org for .gyp change

BUG=161409

Review URL: https://codereview.chromium.org/11343058