<webview>: Cleanup Shim
The shim had both a mutation observer on attributes and getters/setters. This was a bit messy.

The mutation observer on attributes was unnecessary so I got rid of that and added a general
set of getters and setters.

This also fixes some of the issues we were seeing with the partition attribute

Bug=161343
Test=Added test to WebViewTest.Shim to verify that exceptions make it through the shim.

Review URL: https://chromiumcodereview.appspot.com/11412028