FTP: Only issue EPSV command for non-IPv4 connections.
For an IPv4 connection PASV is completely fine. Some servers
become confused by EPSV: they send a success response, but our
connection to data port times out, while with PASV it works fine.

Note that it also might be an issue with firewall between us and the server.
Whatever it is, EPSV confuses it, so we don't use it when it's not necessary.

This also makes the unit test for FTP network transaction test both IPv4
and IPv6 flow.

BUG=131847
TEST=Covered by net_unittests, also see bug.


Review URL: https://chromiumcodereview.appspot.com/11418035