Make browser_tests depend on app_host.
Any test that attempts an install of a platform app would trigger a search for an available app_host.exe and, in its absence, attempt to install it from Chrome's cached installer.

During development, make sure that the development app_host.exe is available as it is inappropriate for the browser_tests to attempt to invoke the system's cached Chrome installer.

BUG=None


Review URL: https://chromiumcodereview.appspot.com/11369252