Tear down AudioPipeReader on host shutdown, to release audio thread.
The AudioPipeReader holds a reference to the audio thread, which
causes it and the host process main thread to remain in-use even
after the host has shutdown, preventing the process from exiting.

This CL works around the problem by releasing the global object
when the host shutdown completes; in future the AudioPipeReader
reference will be moved into the DesktopEnvironmentFactory.

BUG=161373


Review URL: https://chromiumcodereview.appspot.com/11280048