overscroll-aura: Discard synthetic mouse-move events during a trackpad overscroll.
As the overscroll is handled during scroll events from the trackpad, the RWHVA
window is transformed. This transform triggers a synthetic mouse-move event to
be generated (by the aura RootWindow). But this event interferes with the
overscroll gesture. So, ignore such synthetic mouse-move events if an overscroll
gesture is in progress.

BUG=160668

Review URL: https://codereview.chromium.org/11411059