Check before accessing TabSpecificContentSettings
The TSCS' constructor can send notifications that can end up being forwarded to
the ContentSettingImageModel. However, before the constructor returns, the TSCS
is not registered as WebContentsUserData, so we can't get from the WebContents
to the TSCS yet.

I verified that the unit test crashes without the changes to the image model.

BUG=161854
TEST=unit_tests:ContentSettingImageModelTest.NULLTabSpecificContentSettings
R=avi@chromium.org
NOTRY=true


Review URL: https://chromiumcodereview.appspot.com/11415061