Browser Plugin: Simplified BrowserPluginManager
BrowserPluginManager is no longer a singleton.

BrowserPluginManager is now a RenderViewObserver.
All BrowserPluginMsg_* are routed which means they're targeted at a particular RenderView.

This means that more than one BrowserPluginManager can exist per process.

BrowserPluginManager is now refcounted and so it exists as long as a BrowserPlugin or RenderViewImpl
holds a reference to it.

In tests, we inject a BrowserPluginManagerFactory that creates MockBrowserPluginManagers.

BUG=159936
Test=BrowserPluginTest.*, BrowserPluginHostTest.*


Review URL: https://chromiumcodereview.appspot.com/11416074