Sync: Add DeviceInfo's ChangeProcessor
This is the long-awaited change to enable DeviceInfo tracking.

Long ago (r161496) we committed code to support the DeviceInfo type, but
we had to leave this code disabled until the server was ready to support
the new type.  That support has now been added.

This change includes a very special kind of ChangeProcessor named
SyncedDeviceTracker.  It is the only ChangeProcessor that is owned by
the sync thread.  The SyncBackendHost creates and initializes it during
backend intialization and deletes it before the UserShare it references
is destroyed.  As part of its initialization, the SyncedDeviceTracker
will update the DeviceInfo entry for the current device.

The SyncedDeviceTracker could support sending notifications to any
interested listeners when new or update DeviceInfo information arrives,
but this functionality has not been implemented yet.

Also included are:
- Tests for the SyncedDeviceTracker
- A MockTransactionObserver to support these tests
- A cache_guid() accessor on the SyncManager, used to initialize the
  SyncedDeviceTracker
- Remove support for default constructing a DeviceInfo object

BUG=122825


Review URL: https://chromiumcodereview.appspot.com/11360259