Undo all existing overscan settings before updating to a new overscan settings.
If an existing display has overscan setting and SetDisplayOverscan is called for
a different display, the display manager will call OnNativeDisplayChanged(), and
it will re-apply the overscan setting to other displays, so the effect could be
twice or more. We need to undo the existing settings beforehand.

BUG=161097
TEST=confirmed with the repro steps in the bug


Review URL: https://chromiumcodereview.appspot.com/11362259