Revert 168165 - Revert "Revert 167613 - Cache more GL state both service and client side."
This reverts commit 14efc6d257c73271ae0b70d2e71571b335827570.

The issue appears to be that glBlitFramebufferANGLE fails if GL_SCISSOR_TEST
is enabled and the caching code ends up enabling it places it didn't use to
be. That doesn't explain issues on non-Windows machines though. It disable
before glBlitFramebufferANGLE and glBlitFramebufferEXT on all machines
at the moment.

I tried to get the code to fail on Linux NVidia both debug and release
and Mac Retina both debug and release and using integrated and discrete
but no luck.

AFAICT there's no try bots to test Linux Intel or Mac Intel so if
this still causes tests to fail feel free to revert it

TBR=apatrick@chromium.org,brettw@chromium.org
R=shawnsingh@chromium.org
BUG=160370


Review URL: https://chromiumcodereview.appspot.com/11415003

TBR=gman@chromium.org
Review URL: https://codereview.chromium.org/11280110