Introduce time-based caching of encoded data to FFmpegDemuxer.
Previously demuxing was done when a decoder required data. Now that demuxing is performed asynchronously on a separate thread there's a greater risk of not delivering data in time.

The solution is to cache a minimum amount of encoded data per FFmpegDemuxerStream. After collapsing the decoding and pipeline threads together, FFmpegDemuxerStreams will be able to deliver encoded data to decoders without hopping threads.

Running demuxer_bench reports a ~2-3x improvement. Still not as fast as the original code pre-r168350, but significantly better.

Benchmarking method:
# Run once to warm up IO.
$ ./out/Release/demuxer_bench $FILE

# Take average of 20 runs.
$ for i in $(seq 1 20); do ./out/Release/demuxer_bench $FILE; done | awk '/Time:/ { count = count + 1; sum = sum + $2 } END { print (sum / count) }'

BUG=160640

Review URL: https://codereview.chromium.org/11418022