Pass extensions' Content Security Policy to WebKit when using isolated worlds.
http://webkit.org/b/97398 added the ability for script in an isolated world to
bypass the main world's Content Security Policy by setting a new policy on the
isolated world. This new policy is not yet enforced, but that's the end goal.

This CL uses the new API methods that landed in that WebKit patch to associate
a policy with the isolated world created for extensions' content scripts. This
should allow extensions free reign on a website, regardless of it's defined
CSP. A future patch on the WebKit side will allow the isolated world to enforce
the CSP that's passed in, which will give extensions a clearly understandable
policy that's the same everywhere.

BUG=158716


Review URL: https://chromiumcodereview.appspot.com/11366003