[sync] Fix crash seen when PasswordDataTypeController::StartModels fails
In the case where PasswordDataTypeController::StartModels fails due to
the unavailability of the password store, we currently attempt to call
NonFrontendDataTypeController::StartDoneImpl. This is bad, because
StartDoneImpl attempts to run its callback, which is null, in cases where
NonFrontendDataTypeController::StartAssociating has not yet been called.

This patch does the following:
1) Makes PasswordDataTypeController::StartModels simply return false in
cases where the password store is unavailable, which prevents the running
of a null callback.
2) Fixes a bug in ModelAssociationManager::ModelLoadCallback, where if
a model failed to load, the error was not being plumbed back to the caller.
We now call TypeStartCallback with the right error to allow forward progress.
3) Adds the unit test ProfileSyncServicePasswordTest.FailPasswordStoreLoad.

TBR=tim@chromium.org
BUG=146417
TEST=ProfileSyncServicePasswordTest.FailPasswordStoreLoad

Review URL: https://codereview.chromium.org/11414045