Squeeze ninepatch when layer is smaller than bitmap.
This is to address the case where the layer is tiny.  This resulted in a thick shadow because of the overlay.  The optimal behavior for our current use case is be to divide the patches equally and avoid showing the center.

Covered by new test: verifyDrawQuadsForSqueezedLayer

Review URL: https://chromiumcodereview.appspot.com/11416124