IndexedDB: Move quota check to commit phase of transaction.
The "didCompleteTaskEvents" internal API is being removed, and is where quota
checking was performed to abort tasks when a transaction would push the
database over quota (by a rough estimate). Move this check to the commit()
API instead. This means more put()s may succeed than before as the check is
done later, but the tracking is just a "best effort" and in reality those
put() calls could be followed by a bunch of delete() calls.


Review URL: https://chromiumcodereview.appspot.com/11280128