[android_webview] Don't block the IO thread when reading from an InputStream.
This breaks up the functionality in the AndroidStreamReader..Job into
three separate classes, adds native unittests and makes the Job read
the InputStream on a background thread.
This change adds a separate unittestjava folder because the code under
javatests can't be compiled to be a part of a native unittest APK due to
resource dependencies.

TEST=AndroidWebviewTests,android_webview_unittests
BUG=None

Review URL: https://codereview.chromium.org/11363123