It seems QTMonitorImpl::Start() crashed when getting the initial number of input devices in some conner cases.
crash log:
CMIOUnitCreateFromDescription
CMIOUnitCreateFromDescription
+[QTCaptureVDIGDevice _refreshDevices]
+[QTCaptureVDIGDevice devicesWithIOType:]
+[QTCaptureDevice devicesWithIOType:]
+[QTCaptureDevice inputDevices]
device_monitor_mac.mm:43]	content::DeviceMonitorMac::QTMonitorImpl::Start

I guess it is because some racing between chrome is launched and we enumerate the devices in QTMonitorImpl::Start().
A quick fix will be not to enumerate the devices in Start(), though this means that we will have to send one more device changed notification when a device is changed at the first time.

BUG=162324
TEST=manual go to content settings and look at the the device lists.

Review URL: https://chromiumcodereview.appspot.com/11414127