implTransform needs to scale scroll delta with deviceScaleFactor.
This CL fixes a number of issues. First, it fixes the implTransform for non-unit
device scale factors, fixing the issue about not being able to pan the zoom viewport
all the way to the right/bottom margins. It also allows remove the "fudge-factor" used
for hit testing in LayerTreeHost::adjustEventPointForPinchZoom, and finally it
resolves the issue regarding pinch-zoom not handling the anchor point properly
when device scale factor != 1.

BUG=158089


Review URL: https://chromiumcodereview.appspot.com/11414035