Change models.py to use typed class members instead of a list of strings.
This strongly reduces the mismatch when deserializing and surprises that ensue
when a new member is added and is improperly initialized to None without warning.

This basically reproduces the same overall idea that protobuf/django/appengine
implements for serializable data but in a *much* lighter weight implementation.

R=csharp@chromium.org
BUG=


Review URL: https://chromiumcodereview.appspot.com/11414143