Add support for negative times for foreign sessions.
In the ForeignSessionsHandler we compare remote times to our local
system time. This can lead to negative times, and as such we handle the
case correctly.

This CL uses a default-constructor base::TimeDelta() whenever this
happens, instead of passing in a negative time, which caused tests to
crash since they end up in a NOTREACHED() statement.

BUG=154839
TBR=estade@chromium.org

Review URL: https://chromiumcodereview.appspot.com/11415104