Increase timeout in Forwarder2's daemon.cc.
This is the timeout used to connect to the daemon's Unix Domain Socket. In case
the daemon was just spawned it can easily take a few hundred milliseconds on
the device before it binds the socket.

The same timeout is now also used when killing the daemon.


Review URL: https://chromiumcodereview.appspot.com/11418151