Fixing the hunspell convert_dict tool to replace the affixes from the .dic file with those from the .dic_delta file.
As it stands right now, they are added to the set, but do not make a difference in the actual interpretation. From testing, there is only ever one indexed affix per word so replacing it with the most recent (i.e., the one from the .dic_delta file) is what we want.

The understanding being that if you want to change the affix, you would just add the word to the .dic_delta file with the new affixes you want.


BUG=160847

TBR=sky@chromium.org

Review URL: https://chromiumcodereview.appspot.com/11419044