google_apis: Remove a hard-coded URL hack from HttpRequestParser
"http://localhost/" was hard-coded in HttpRequestParser, but this was wrong
as the server runs at a different host name and a port.

Along the way, add a test for HttpServer::RegisterDefaultResponse, which was
missing. Also add the request URL to warning messages for 404 requests.

BUG=none
TEST=none

Review URL: https://codereview.chromium.org/11280156