Add asserts that should help in diagnosing the crashes being observed.
Change many DCHECKs to CHECKs, so that we can pinpoint where things to wrong
in crash reports.

Rename deleteUnlinkedEvictedBackings to deleteAllEvictedBackings, and make
it assert that all backings are unlinked. This was the only way the call
was used, so this just makes the actual behavior more clear.

Add additional CHECKs to ensure that we don't have any dangling pointers to
backings. Move the unlinking of evicted backings to the very beginning of
beginFrame, so that we can assert that there is an exactly one-to-one
relationship between adding resources to a list to pass to the main thread,
and the main thread unlinking the resources in that list (note that this
entire process may happen multiple times though, until a commit completes).

This very strictly enforces the pattern of: scheduledActionBeginFrame
is always followed by exactly one call to beginFrame, which may be followed
by zero or one calls to scheduledActionCommit, and only after this cycle is
complete may scheduledActionBeginFrame be called again.

Once the issue at hand has been dealt with, we may want to change some of
the CHECKs back to DCHECKs and ensconce the main-thread-passing logic in a
ifdef-DEBUG, if there is believed to be a performance penalty.

BUG=161676

Review URL: https://chromiumcodereview.appspot.com/11348239