Have ExtensionService notify ThemeService directly for themes
Save the theme ID if the service has not yet been created, as creating
the service causes a startup regression.

This is so that the ThemeService correctly picks up the new theme
even if it hasn't been created yet.

Also clean up the various getters/setters a bit.

BUG=161203


Review URL: https://chromiumcodereview.appspot.com/11348238