Bundle resources into the unit_test target on iOS
On iOS, all resources used in tests must be bundled into the unit test
target's app bundle (since the Mac approach of looking next to the binary
in the build directory doesn't work since the app is copied to a simulator
or device before being run).

This creates a gypi that can be included in any target that needs pak
files bundled on iOS (unit_tests, higher-level tests, the app, etc.), and
uses it in unit_tests.

BUG=None


Review URL: https://chromiumcodereview.appspot.com/11412069