Do not remove the notification item if the click happens on the card in popup.
Removing items in popup seems unexpected, because the user won't see it anymore.  Probably better to keep it in popup, but remove it in the message center.  Anyway the user can remove it by clicking "x" button.  Also this CL closes the popup when clicked instead.

I think it also should raise the window (and the tab) of the notification for the click.  It'll be done in another CL.

Review URL: https://chromiumcodereview.appspot.com/11362217