Fix parsing limitation in gsdview.
xml.sax.handler's apparently needs to support multiple content areas in a row
for a give element. (Actually it says this in the doc... sigh).

This url:
http://gsdview.appspot.com/chromium-browser-official/?marker=chromium-21.0.1168.0.tar.bz%40
happened to tickle a case in which whatever internal buffer xml.sax is using
fell mid-element causing the content handler to be called twice, once with each
half of the text. This was then causing other parsing to blow up as we expected
only to be setting (not appending) to the current element.

Setting a default and appending.

BUG=None
TEST=None
R=phajdan@google.com
Review URL: https://codereview.chromium.org/11413203