Undo all existing overscan settings before updating to a new overscan settings (2nd)
This is 2nd patch of crrev.com/168974, which was reverted due to win_aura test failure.
I noticed that this is essentially a bug of win-aura, so I disabled the broken test for windows temporarily.

Original description:
If an existing display has overscan setting and SetDisplayOverscan is called for
a different display, the display manager will call OnNativeDisplayChanged(), and
it will re-apply the overscan setting to other displays, so the effect could be
twice or more. We need to undo the existing settings beforehand.

BUG=161097

Review URL: https://chromiumcodereview.appspot.com/11417121