ash: Do not snap when touch-resizing a window, but mostly code cleanup.
Do not snap a window if a gesture-drag is resizing it (as opposed to moving it).
Also, use the API functions to get/set the restore bounds, instead of using the
key directly. This makes the code a bit cleaner, and makes it easier to track
who/when sets the restore bounds on windows.

BUG=162510,151267

Review URL: https://codereview.chromium.org/11418211