InstantExtended: Restrict allowed heights.
Only honor the height from show() in certain situations. Otherwise,
show at 100% height. This is to avoid having to pass to the searchbox
Javascript the state that lets it deduce it is now on a page where
it ought to show the dropdown not at full height.

BUG=


Review URL: https://chromiumcodereview.appspot.com/11299221