Fix tautological compares in font_atlas.cc.
This fixes the following warnings from a recent version of Clang:

cc/font_atlas.cc:62:43: error: comparison of constant 128 with
expression of type 'char' is always true
[-Werror,-Wtautological-constant-out-of-range-compare]

cc/font_atlas.cc:45:39: error: comparison of constant 128 with
expression of type 'const char' is always true
[-Werror,-Wtautological-constant-out-of-range-compare]

The problem is that the char type is signed on this platform, and
therefore always < 128. The code really wants to work with the unsigned
values here.

BUG=163104


Review URL: https://chromiumcodereview.appspot.com/11416239