cc: Add a test to verify DelegatedRendererLayer owns a RenderSurface when needed.
In the test, the layer is holding quads from a tree that consisted of multiple
layers, and the quads overlap. When the DelegatedRendererLayer has opacity 1,
it does not need to own a surface here. But when it has opacity < 1, then it
does.

R=piman
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/11421089