Remove remoting.HostController.state()
In future, when we remove NPAPI we'll be getting host state asynchronously.
This change removes state() method from HostController class. Now the
getLocalHostStateAndId() should be used to get host state. There is still
isInstalled() used by the HostSetupDialog, but it the future it will need
to be removed as well.

Review URL: https://codereview.chromium.org/11418207