Support creating virtual contexts.
This creates virtual contexts within a GLShareGroup, when
'--enable-virtual-gl-contexts' is given on the command line.
It requires the shared real GL context, which is created
from the default offscreen surface configuration, to be compatible
with any surface that is made current with a virtual context.

BUG=160710


Review URL: https://chromiumcodereview.appspot.com/11308151