bookmarks: Pass BrowserContext to ToggleWhenVisible() function.
We are switching away from Profile, and content::BrowserContext is preferred
over it.

This patch was moves it to ui/bookmarks, because it's only used by c/b/ui,
and renames it to ToggleBookmarkBarWhenVisible() so it's more clear what is being
toggled for the consumers.

BUG=144783
R=sky@chromium.org

Review URL: https://codereview.chromium.org/11411228