Refactor the test to be more testable and add a regression test.
The test actually tests that the code is broken since 'try job rietveld' should
be SUCCEEDED. The fix will (hopefully) be in a follow up change. I'm creating
this CL to make the fix more obvious (or postpone the fix, depending on the
point of view).

R=csharp@chromium.org
BUG=


Review URL: https://chromiumcodereview.appspot.com/11316226