[NaCl SDK] remove --toolchain option from create_nmf.
This simplifies the create_nmf invocations in our example
Makefiles.

The --toolchain option was only needed to determin if an
executable is dynamic or not (e.g. if it was built with
glibc toolchain or not).  With this change create_nmf now
knows enough about the elf headers to determine this
on its own.

The -t/--toolchain is now deprecated (but not removed
in case anyone was using it).

R=binji
BUG=


Review URL: https://chromiumcodereview.appspot.com/11348287