CrOS: Add hotkey for disabling the GPU watchdog timer.
This works around the fact that switching to VT1 on CrOS prevents swaps from clearing, which hangs the GPU process, which causes the GPU process to get killed. Sometimes in order to diagnose a problem, we want to switch to VT1. This hotkey allows a developer to do that.

TEST=by hand; Ctrl-Shift-Alt-G; then go about://gpuhang; UI should freeze and not unstick after 10 seconds
BUG=chromium-os:28253


Review URL: https://chromiumcodereview.appspot.com/11348099