Reenable notification tests on win_aura, and move to interactive_ui_tests
Disabled at http://crrev.com/170051

These tests eventually refer to CheckIdleStateIsLocked, which blocks
notifications from showing when the desktop is locked, or switched to
a non-default desktop. So, move them to interactive_ui_tests where we're
guaranteed to be running as a logged in user.

R=jam@chromium.org
BUG=154081

Review URL: https://codereview.chromium.org/11316252