Texture Draw Calls Coalescing
This patch batches multiple calls to DrawQuad inside GlRenderer when the calls draw textured quads that have the same texture.  These quads differ only in transform for both position and UV.  This patch extends the vertex shader used for drawing textures to use one of 8 matrices and UV transforms.  By batching up to 8 DrawQuads into a single draw quad, we reduce the number of times many different OpenGL ES state calls are made.

The implementation maintains a cache of up to 8 quads that contain the same texture and drawing parameters (e.g. opacity).  If a 9th textured quad with the same parameters or a different kind of quad is drawn or any GL state (e.g. scissor rect) is changed or the new finishDrawingQuadList() function is called then the the quads are drawn in a single call and the cache is flushed.

BUG=161372

Review URL: https://chromiumcodereview.appspot.com/11415161