[telemetry] Fix smoothScrollBy scroll distance calculation.
The smoothness benchmark mean frame times were off because they would attempt to smoothScrollBy to the scrollHeight, but you need to subtract the viewport height to get the correct scrolling distance. Because smoothScrollBy doesn't stop when you hit the bottom of the page, the duration of the scroll gesture was longer than it should be. This made the numbers worse for shorter pages.

BUG=162922
TEST=Run a smoothness_benchmark.

Review URL: https://codereview.chromium.org/11412266