gpu: Defer reads from default framebuffer when needed.
On some platforms, we want to unschedule after a swap until the ack, but the
DeferDraws mechanism allows us to delay the unschedule until we actually need to
access the back buffer. Up to now, only commands that *write* to the framebuffer
were considered, however we also need to consider *reads*. This fixes that.

BUG=None


Review URL: https://chromiumcodereview.appspot.com/11412262