Add public accounts to UserManager
This CL extends the UserManager to handle public accounts defined through
policy. User pods are dynamically added and removed when the list of
public accounts in policy changes. Any data belonging to obsolete accounts
is also removed, taking care not to remove it prematurely if a user is
currently logged into the account.

The CL also makes the user list handling more robust by checking for
duplicate entries in the user list prefs and logging these as errors.

The pods added for public accounts are not functional yet. The login flow
for public accounts will be the topic of another CL.

BUG=158509


Review URL: https://chromiumcodereview.appspot.com/11419184