Use textInputInfo instead of textInputType in getTextInputType
For the given repro case, textInputType seems to get stuck in the
previous input type whereas textInputInfo is checking whether
an editor exists for the focused frame and exiting early with
with TextInputType set to None. Switching to using textInputInfo
in getTextInputType, as the comments in the header suggests the use of
textInputType by itself as an API should be deprecated anyway.

BUG=162831


Review URL: https://chromiumcodereview.appspot.com/11419218