[Android WebView] AwContentsClient.shouldCreate window callback part 2.
Part 1: https://chromiumcodereview.appspot.com/11362183/

The second part of this change implements the "true" path,
i.e. the user provides us with a new AwContents to host
the popup window. This arrives asynchronously.

While we are waiting for the reply from the user, we defer
loading of the popup contents using a resource throttle.

When we receive the callback we replace the ContentViewCore that
is contained in the new AwContents with one that wraps the WebContents
for the popup. This triggers an update to the AwIoThreadClient for
that AwContents, which resumes the deferred load for the popup.

Android only change and android bots green
NOTRY=TRUE

Review URL: https://chromiumcodereview.appspot.com/11348075