Add [content, media, net, ui]_unittests for iOS to main and try waterfalls.
This CL adds several unit tests to the mac and try waterfalls as well as to
the gatekeeper.

These tests have been running on the iOS simulator builder on the fyi waterfall
for a week or so.

The CL for adding these to the commit queue is at https://codereview.chromium.org/11360234

BUG=160788

Review URL: https://chromiumcodereview.appspot.com/11359190