Send 'IP address changed' events on Android when connection type changes.
NetworkChangeNotifierAndroid was only firing 'network connection changed'
events although many clients subscribe to 'IP address changed events'.
One of them is ClientSocketPoolBaseHelper which invalidates its socket pool
whenever the IP address changes. The lack of this notification on Android could
have led to an invalid socket being used.

This is a first approach. Later on this can be improved to only fire the
notification when the IP address actually changed.

This issue was found as part of the following bug investigation although this
CL doesn't claim to fix the bug.

BUG=149231

Review URL: https://codereview.chromium.org/11411287