Fix constrained window position for inactive tab
If a constrained window was added to an inactive tab it would be incorrectly positioned.

The problem was that an inactive tab view is removed from the browser window. This meant that the constrained window position coudln't be calculated.

When the tab became activated again we weren't updating the sheet position.

This CL changes the code to correclty update the sheet position when the tab is activated.

BUG=163841
TEST=Verified that the bug can no longer be reproduced. Verified that ConstrainedWindowSheetControllerTest.AddToInactiveTab failed without my change and passes with it.


Review URL: https://chromiumcodereview.appspot.com/11416309