Customize user details in ash system bubble for public account mode
This CL customizes the user details shown at the top of the ash system
bubble when logged into a public account. The regular display of user name
and e-mail is replaced by a text that informs the user this is a public
account and names the domain managing the account. The implementation
follows the UI mock in the linked bug.

BUG=152928


Review URL: https://chromiumcodereview.appspot.com/11377133