cc: Give ownership of render passes to the renderer when drawing a frame.
Instead of giving const pointers to the renderer, we should provide it with
non-const pointers, and expect it to consume the render passes that we provide
it.

In ubercomp, the delegating renderer will pass ownership of the passes to the
CompositorFrame. 

Covered by existing tests, no change in behaviour.

BUG=152337
R=enne,piman
NOTRY=true

Review URL: https://chromiumcodereview.appspot.com/11419269