Video capture implementation using the Media Foundation API. This replaces the DirectShow based implementation for Vista,Win7 and Win8. DirectShow isn't supported in metro mode so once we've landed this, we can revert the workaround we have in place for Win8.
The CapabilityList class is mostly moved code from the DS implementation.
The difference is that I'm not using a std::map<> since that wasn't
really necessary and instead adding one member (via inheritance) to
the capability struct on Windows that holds the stream id that is needed.

BUG=140545
TEST=All video capture tests in media_unittests now test this new implementation (except on XP).

Review URL: https://codereview.chromium.org/11419200