Query the network delegate for cookies even if there is no cookie monster
On iOS, there is a platform specific implementation of the cookie store which is not a CookieMonster.
However, the network delegate is currently only called if the cookie store is a CookieMonster, and thus cookie settings are bypassed.
This CL calls the network delegate even if the cookie store is not a CookieMonster.

BUG=145954

Review URL: https://chromiumcodereview.appspot.com/11417148