Fixed issue with miscalculated aria-posinset attribute's value.
This bug was caused by returning a string as an index from list_selection_model.js in an Event argument of a change event. In the list.js, such index stored as string was incremented, which caused concatenation instead of incrementation. This patch converts these values to number to avoid such problems.

TEST=Enable ChromeVox, enter Files.app (CTRL-M), navigate among files. Correct item number should be spoken, eg. 2/x, 3/x.
BUG=164061


Review URL: https://chromiumcodereview.appspot.com/11421214