Depend on stored sync session GUID for Android.
This changes the Android port of Chrome to depend on the pref
kSyncSessionsGUID to be set before the SessionModelAssociator is
initialized.

We now also force-removes any foreign sessions that are using the
current client's GUID as a session tag.

This also lead to the removal of the Android specific session utils.

BUG=157419


Review URL: https://chromiumcodereview.appspot.com/11414013