Removes abuse of the word 'hash' and rename a fair number of variables.
Use 'hash' only when it makes sense. This CL requires a wipe of the whole server.

This CL makes it potentially more possible to change the hashing algorithm in
the future, by keeping each hash algorithm in its own namespace. It's not
completely done; all hash are currently required to be 20 bytes, but it's is
progress towards that goal.

R=csharp@chromium.org
BUG=


Review URL: https://chromiumcodereview.appspot.com/11280293