If a new chrome window becomes visible while the current metro chrome window is snapped, the new window should be snapped.
Currently the new chrome window which becomes visible ends up with the wrong size when metro chrome is snapped. We should ensure
that the new chrome window is also snapped and the IDC_METRO_SNAP_ENABLE message is sent to it. This ensures that the behavior remains
consistent, i.e the omnibox remains invisible, etc.

BUG=163458
R=robertshield
Review URL: https://codereview.chromium.org/11442004