CQ: Look for stale pending job requests that are not queued anymore.
This was a significant source of CQ hangs; a Try Job was requested on Rietveld
but didn't go up to the Try Server. This ensures the connectivity is up between
the TS and Rietveld. It gives a 5 minutes delay before enforcing the check, to
reduce the number of requests to <builder>/pendingBuilds and to accomodate with
propagation delays.

R=csharp@chromium.org
BUG=161259


Review URL: https://chromiumcodereview.appspot.com/11455003