Make page transition type to InterceptNavigationResourceThrottlethrottle
The intercept navigation resource throttle needs to be able to
inspect the page transition type associated with a request, as
different behaviour may be desired depending on it. For example,
in Chrome on Android, we do not want to intercept requests
resulting from backwards / forwards history navigation.

BUG=164194

Review URL: https://chromiumcodereview.appspot.com/11415252