Organize internal properties of cc/ layer types
One useful way to classify the various properties in cc/ layer data types is:
(1) properties given by the user of the cc APIs, and (2) properties that are
computed inside of cc code that are needed to correctly draw the layers.

This patch organizes properties internally so that as we develop cc code, we
are forced to be aware of whether a property is "given" or "computed".

Review URL: https://codereview.chromium.org/11280263