Improve HostResource comments for use in plugins.
The comments in host_resource.h allude to a transformation that must
happen for a PP_Resource to be used in a plugin, but doesn't specify
what it is. Added a hint so it's easier to find what to use.

Also corrected a spelling error.

BUG=


Review URL: https://chromiumcodereview.appspot.com/11437017