Fix showing constrained window for uninitialized tabs
If a tab is added without being shown then the associated tab view for the web content will not be created yet.

When a constrained window is added to that tab the window never gets shown.

To fix this I've changed ConstrainedWindowMac2 to reshow the window once the web contents becomes visible.

Note, this change was reverted because ConstrainedWindowMacTest.ShowInUninitializedTab was failing. The code that was causing the failure was deleted in r171328 so it's safe to check this in again. (Old commit revision r170795).

BUG=163778
TEST=Reproduced bug 163778 and verified that my patch fixes it.
Verified that ConstrainedWindowMacTest.ShowInUninitializedTab fails without my patch and passes with it.

Review URL: https://chromiumcodereview.appspot.com/11419277