Use Search-F# to generate Fkey keycodes.
This is in addition to Search-#. On a Chromebook, it would only let you
generate F1 to F10. But on an external keyboard, it could let you generate
F1 to F12, and is more intuitive.

This also matches typical Fn-key behaviour on other keyboards with such a key.

R=yusukes@chromium.org,mazda@chromium.org,derat@chromium.org
BUG=164216

Review URL: https://chromiumcodereview.appspot.com/11437039