[cc] Move root layer pointer to LayerTreeImpl
This is the first of a series of patches to move tree-specific state from
LayerTreeHostImpl to LayerTreeImpl. The LayerTreeHostImpl has always been
imagined to be a "View" aka LayerTreeView(Impl) and will keep ahold of stuff
that is view-specific.  The LayerTreeImpl, in constrast, will store state that
is specific to a specific tree version.

BUG=155209
R=enne


Review URL: https://chromiumcodereview.appspot.com/11450015