Add singleton argument to app.window.create().
If a window is given an id, trying to create a window with an id that is already
used will cause the existing window to be activated instead of a new window to
be created. You can override this behavior by passing singleton: false.

BUG=159217


Review URL: https://chromiumcodereview.appspot.com/11419295