Omnibox: Revise HistoryQuickProvider Code for URL-What-You-Typed Goes First (redux)
This is like
https://codereview.chromium.org/11415265/
with one minor change.  That change was reverted because it caused a failure in the perf test for static initializers.

Numbers such as
// static
const int foo = 1400 + kMaxMatches;
are not compile time constants (because kMaxMatches isn't in the same .cc file) and hence cause this initialization to happen at runtime at startup.  This is prohibited.

I tried putting the definitions for these in the .h but the linker didn't like that.

Instead, I thought some more and realized that these scores doesn't need to depend on kMaxMatches.  The only reason that's in there is historic.  If we change kMaxMatches, there's no reason we should change the value of these scores.  Hence, I simply made them all constant with kMaxMatches.

I'm sorry I can't attach this change to the previous one.  git cl upload won't let me.

---
original changelist description for reference:
I noticed that the code in HQP to guarantee thatURL-what-you-typed goes first (when it's available) is too aggressive. For instance, when URL-what-you-typed is a visited URL, it gets a score of 1413 currently. But HQP makes the scores of its results <1200. This means that other provider's results (such as search suggestions) can sneak in in the middle. In this example, after this change HQP realizes the URL-what-you-typed result will have score 1413, and so it simply makes sure its results have scores less than 1413.

This is a step in making HQP be able to replace HUP.

In the process of making this change, I had to expose HUP's relevancy scores. I made them a public static const int.

I know these scores are used in scored_history_match in HUP-like-scoring mode. Hence I changed the code in scored_history_match to use these newly exposed constants.


BUG=


Review URL: https://chromiumcodereview.appspot.com/11443032